RasPi/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
Pillow==7.0.0
requests==2.23.0
spidev==3.4
urllib3==1.25.8
